Move model crate stubs into defaults
