+++ nautilus_core/model/src/accounts/any.rs
@@ -145,13 +145,6 @@ impl From<AccountState> for AccountAny {
     }
 }
 
-impl Default for AccountAny {
-    /// Creates a new default [`AccountAny`] instance.
-    fn default() -> Self {
-        AccountAny::Cash(CashAccount::default())
-    }
-}
-
 impl PartialEq for AccountAny {
     fn eq(&self, other: &Self) -> bool {
         self.id() == other.id()
