+++ nautilus_core/model/src/accounts/cash.rs
@@ -26,10 +26,7 @@ use crate::{
     accounts::base::{Account, BaseAccount},
     enums::{AccountType, LiquiditySide, OrderSide},
     events::{AccountState, OrderFilled},
-    identifiers::{
-        stubs::{account_id, uuid4},
-        AccountId,
-    },
+    identifiers::AccountId,
     instruments::InstrumentAny,
     position::Position,
     types::{AccountBalance, Currency, Money, Price, Quantity},
@@ -246,29 +243,6 @@ impl Display for CashAccount {
                 |base_currency| format!("{}", base_currency.code)
             ),
         )
-    }
-}
-
-impl Default for CashAccount {
-    /// Creates a new default [`CashAccount`] instance.
-    fn default() -> Self {
-        // million dollar account
-        let init_event = AccountState::new(
-            account_id(),
-            AccountType::Cash,
-            vec![AccountBalance::new(
-                Money::from("1000000 USD"),
-                Money::from("0 USD"),
-                Money::from("1000000 USD"),
-            )],
-            vec![],
-            true,
-            uuid4(),
-            0.into(),
-            0.into(),
-            Some(Currency::USD()),
-        );
-        Self::new(init_event, false)
     }
 }
