+++ nautilus_core/model/src/identifiers/mod.rs
@@ -24,6 +24,7 @@ pub mod account_id;
 pub mod client_id;
 pub mod client_order_id;
 pub mod component_id;
+pub mod default;
 pub mod exec_algorithm_id;
 pub mod instrument_id;
 pub mod order_list_id;
